Resolve edit hashes to applications

+++ util/impl_cleanup.sh
@@ -3,11 +3,17 @@
 SCRIPT_DIR="$(cd "$(dirname "$0")" && pwd)"
 cd "$SCRIPT_DIR"
 
-sudo systemctl stop server_$1
-sleep 2
-sudo systemctl disable server_$1
-sudo rm /etc/systemd/system/server_${1}.service
-sudo rm -rf ~/servers/$1
+if [ -f "/etc/systemd/system/server_$1" ]; then
+  sudo systemctl stop server_$1
+  sleep 2
+  sudo systemctl disable server_$1
+  sudo rm /etc/systemd/system/server_${1}.service
+fi
+
+if [ -d "~/servers/$1" ]; then
+  sudo rm -rf ~/servers/$1
+fi
+
 sudo rm /etc/nginx/sites-available/$1.app /etc/nginx/sites-enabled/$1.app
 sudo nginx -s reload
 sps-client application delete --name $1
